hb/scripts/requirements.txt
sqlalchemy>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.3
h3>=3.7.4
shapely>=1.8.0
geopandas>=0.10.2
pandas>=1.3.5 